+++ src/pages/TokenDetails/index.tsx
@@ -1,150 +1,57 @@
-import { Currency, Token } from '@uniswap/sdk-core'
-import { PageName } from 'analytics/constants'
-import { Trace } from 'analytics/Trace'
-import { AboutSection } from 'components/Tokens/TokenDetails/About'
-import AddressSection from 'components/Tokens/TokenDetails/AddressSection'
-import BalanceSummary from 'components/Tokens/TokenDetails/BalanceSummary'
-import { BreadcrumbNavLink } from 'components/Tokens/TokenDetails/BreadcrumbNavLink'
-import ChartSection from 'components/Tokens/TokenDetails/ChartSection'
-import MobileBalanceSummaryFooter from 'components/Tokens/TokenDetails/MobileBalanceSummaryFooter'
-import TokenDetailsSkeleton, {
-  Hr,
-  LeftPanel,
-  RightPanel,
-  TokenDetailsLayout,
-} from 'components/Tokens/TokenDetails/Skeleton'
-import StatsSection from 'components/Tokens/TokenDetails/StatsSection'
-import TokenSafetyMessage from 'components/TokenSafety/TokenSafetyMessage'
-import TokenSafetyModal from 'components/TokenSafety/TokenSafetyModal'
-import Widget from 'components/Widget'
-import { DEFAULT_ERC20_DECIMALS, NATIVE_CHAIN_ID, nativeOnChain } from 'constants/tokens'
-import { checkWarning } from 'constants/tokenSafety'
-import { Chain } from 'graphql/data/__generated__/TokenQuery.graphql'
-import { QueryToken, useTokenQuery } from 'graphql/data/Token'
-import { useTokenPriceQuery } from 'graphql/data/TokenPrice'
-import { CHAIN_NAME_TO_CHAIN_ID, validateUrlChainParam } from 'graphql/data/util'
-import { useIsUserAddedTokenOnChain } from 'hooks/Tokens'
-import { useOnGlobalChainSwitch } from 'hooks/useGlobalChainSwitch'
-import { useCallback, useMemo, useState, useTransition } from 'react'
-import { ArrowLeft } from 'react-feather'
-import { useNavigate, useParams } from 'react-router-dom'
+import { filterTimeAtom } from 'components/Tokens/state'
+import TokenDetails from 'components/Tokens/TokenDetails'
+import { TokenDetailsPageSkeleton } from 'components/Tokens/TokenDetails/Skeleton'
+import { NATIVE_CHAIN_ID, nativeOnChain } from 'constants/tokens'
+import { TokenQuery, tokenQuery } from 'graphql/data/Token'
+import { TokenPriceQuery, tokenPriceQuery } from 'graphql/data/TokenPrice'
+import { CHAIN_NAME_TO_CHAIN_ID, TimePeriod, toHistoryDuration, validateUrlChainParam } from 'graphql/data/util'
+import { useAtomValue } from 'jotai/utils'
+import { Suspense, useCallback, useEffect, useMemo } from 'react'
+import { useQueryLoader } from 'react-relay'
+import { useParams } from 'react-router-dom'
 
-export default function TokenDetails() {
+export default function TokenDetailsPage() {
   const { tokenAddress, chainName } = useParams<{ tokenAddress?: string; chainName?: string }>()
   const chain = validateUrlChainParam(chainName)
   const pageChainId = CHAIN_NAME_TO_CHAIN_ID[chain]
-  const nativeCurrency = nativeOnChain(pageChainId)
   const isNative = tokenAddress === NATIVE_CHAIN_ID
-  const tokenQueryData = useTokenQuery(isNative ? nativeCurrency.wrapped.address : tokenAddress ?? '', chain)
-  const prices = useTokenPriceQuery(isNative ? nativeCurrency.wrapped.address : tokenAddress ?? '', chain)
-  const token = useMemo(() => {
-    if (!tokenAddress) return undefined
-    if (isNative) return nativeCurrency
-    if (tokenQueryData) return new QueryToken(tokenQueryData)
-    return new Token(pageChainId, tokenAddress, DEFAULT_ERC20_DECIMALS)
-  }, [isNative, nativeCurrency, pageChainId, tokenAddress, tokenQueryData])
+  const timePeriod = useAtomValue(filterTimeAtom)
+  const [contract, duration] = useMemo(
+    () => [
+      { address: isNative ? nativeOnChain(pageChainId).wrapped.address : tokenAddress ?? '', chain },
+      toHistoryDuration(timePeriod),
+    ],
+    [chain, isNative, pageChainId, timePeriod, tokenAddress]
+  )
 
-  const tokenWarning = tokenAddress ? checkWarning(tokenAddress) : null
-  const isBlockedToken = tokenWarning?.canProceed === false
+  const [tokenQueryReference, loadTokenQuery] = useQueryLoader<TokenQuery>(tokenQuery)
+  const [priceQueryReference, loadPriceQuery] = useQueryLoader<TokenPriceQuery>(tokenPriceQuery)
 
-  const navigate = useNavigate()
-  // Wrapping navigate in a transition prevents Suspense from unnecessarily showing fallbacks again.
-  const [isPending, startTransition] = useTransition()
-  const navigateToTokenForChain = useCallback(
-    (chain: Chain) => {
-      const chainName = chain.toLowerCase()
-      const token = tokenQueryData?.project?.tokens.find((token) => token.chain === chain && token.address)
-      const address = isNative ? NATIVE_CHAIN_ID : token?.address
-      if (!address) return
-      startTransition(() => navigate(`/tokens/${chainName}/${address}`))
+  useEffect(() => {
+    loadTokenQuery({ contract })
+    loadPriceQuery({ contract, duration })
+  }, [contract, duration, loadPriceQuery, loadTokenQuery, timePeriod])
+
+  const refetchTokenPrices = useCallback(
+    (t: TimePeriod) => {
+      loadPriceQuery({ contract, duration: toHistoryDuration(t) })
     },
-    [isNative, navigate, tokenQueryData?.project?.tokens]
-  )
-  useOnGlobalChainSwitch(navigateToTokenForChain)
-  const navigateToWidgetSelectedToken = useCallback(
-    (token: Currency) => {
-      const address = token.isNative ? NATIVE_CHAIN_ID : token.address
-      startTransition(() => navigate(`/tokens/${chainName}/${address}`))
-    },
-    [chainName, navigate]
+    [contract, loadPriceQuery]
   )
 
-  const [continueSwap, setContinueSwap] = useState<{ resolve: (value: boolean | PromiseLike<boolean>) => void }>()
-
-  // Show token safety modal if Swap-reviewing a warning token, at all times if the current token is blocked
-  const shouldShowSpeedbump = !useIsUserAddedTokenOnChain(tokenAddress, pageChainId) && tokenWarning !== null
-  const onReviewSwapClick = useCallback(
-    () => new Promise<boolean>((resolve) => (shouldShowSpeedbump ? setContinueSwap({ resolve }) : resolve(true))),
-    [shouldShowSpeedbump]
-  )
-
-  const onResolveSwap = useCallback(
-    (value: boolean) => {
-      continueSwap?.resolve(value)
-      setContinueSwap(undefined)
-    },
-    [continueSwap, setContinueSwap]
-  )
+  if (!tokenQueryReference) {
+    return <TokenDetailsPageSkeleton />
+  }
 
   return (
-    <Trace page={PageName.TOKEN_DETAILS_PAGE} properties={{ tokenAddress, tokenName: chainName }} shouldLogImpression>
-      <TokenDetailsLayout>
-        {tokenQueryData && !isPending ? (
-          <LeftPanel>
-            <BreadcrumbNavLink to={`/tokens/${chainName}`}>
-              <ArrowLeft size={14} /> Tokens
-            </BreadcrumbNavLink>
-            <ChartSection
-              token={tokenQueryData}
-              currency={token}
-              nativeCurrency={isNative ? nativeCurrency : undefined}
-              prices={prices}
-            />
-            <StatsSection
-              TVL={tokenQueryData.market?.totalValueLocked?.value}
-              volume24H={tokenQueryData.market?.volume24H?.value}
-              priceHigh52W={tokenQueryData.market?.priceHigh52W?.value}
-              priceLow52W={tokenQueryData.market?.priceLow52W?.value}
-            />
-            {!isNative && (
-              <>
-                <Hr />
-                <AboutSection
-                  address={tokenQueryData.address ?? ''}
-                  description={tokenQueryData.project?.description}
-                  homepageUrl={tokenQueryData.project?.homepageUrl}
-                  twitterName={tokenQueryData.project?.twitterName}
-                />
-                <AddressSection address={tokenQueryData.address ?? ''} />
-              </>
-            )}
-          </LeftPanel>
-        ) : (
-          <TokenDetailsSkeleton />
-        )}
-
-        <RightPanel>
-          <Widget
-            token={token ?? nativeCurrency}
-            onTokenChange={navigateToWidgetSelectedToken}
-            onReviewSwapClick={onReviewSwapClick}
-          />
-          {tokenWarning && <TokenSafetyMessage tokenAddress={tokenAddress ?? ''} warning={tokenWarning} />}
-          {token && <BalanceSummary token={token} />}
-        </RightPanel>
-        {token && <MobileBalanceSummaryFooter token={token} />}
-
-        {tokenAddress && (
-          <TokenSafetyModal
-            isOpen={isBlockedToken || !!continueSwap}
-            tokenAddress={tokenAddress}
-            onContinue={() => onResolveSwap(true)}
-            onBlocked={() => navigate(-1)}
-            onCancel={() => onResolveSwap(false)}
-            showCancel={true}
-          />
-        )}
-      </TokenDetailsLayout>
-    </Trace>
+    <Suspense fallback={<TokenDetailsPageSkeleton />}>
+      <TokenDetails
+        tokenAddress={tokenAddress}
+        chain={chain}
+        tokenQueryReference={tokenQueryReference}
+        priceQueryReference={priceQueryReference}
+        refetchTokenPrices={refetchTokenPrices}
+      />
+    </Suspense>
   )
 }
